User password is securely harvested
I had to add a dependency on a third party library.
I reviewed the source and tested the implementation
and don't see any security concerns. Usernames and
passwords are stored in an encrypted shared preferences
file. The encryption secret is the users password, so
the information is reset automatically every time it can't
be read (i.e. the password changes).