Added expandable groupings for settings
I can't believe this worked nearly the first time through, but
settings are now in an expandable list. The default is to have
the list expanded, with collapse available if necessary. If there
is a way to group hidden, or mark them as hidden/inactive in the
app this could be used to hide them.