Refactored login flow to check session first
Session verification and subsequent login was pulled into the
TracsClient class where it should've been in the first place. Any
activity used should verify the session with TracsClient prior to
loading potentially sensitive information. If the session is not
valid then the user is routed home to the CAS login screen.