Added swipe to dismiss notifications
The animation and onclick functionality are implemented. This was
ten times harder than it should be; animations have to be manually
performed and gestures detected. Still need to implement exit
velocity when animating the notification and draw some sort of border
or separator on the notifications. Also needs to mark the notification
as cleared, but that is inconvenient for testing.