Refactored how requests are tracked for cancellation
Most of stack overflow says you should tag volley requests with
a string. As it turns out you should use an object, and that object
has to be used again when you cancel the request. If you use the same
string but it's not the same object the request will not be cancelled.