Notification icon and badge needed to have button behavior
Since I'm using a custom text view as a menu button I had to
establish a bit of custom behavior for the view so it would behave
like a button. This happens to be the button tied to login status
so it was more involved than a normal button should be.