Fixed notifications indefinitely hanging bug
If the app was supplied with a notification that it could not
retrieve (due to 403 or 500 responses) the app would indefinitely
wait for the remaining notifications to be added to the list.
I now remove any notification from the list to be retrieved if
the response produces an error code.