Fixed hanging Notifications screen from push notification
Push notifications launch the MainActivity which hooks into the
notifications activity. The use of finish() was still allowing
verifySession to be called, which would call the callback on
completion. The notifications activity was replacing the callback
before the request came back, and causing its own callback to
be called twice despite only making one request.