Fixed bug preventing navigation to announcements from notification
Android has an additional lifecycle event that appears to be on the
super-secret menu. The method onNewIntent is called when you launch
an activity that has already been created. In this case the old
intent still exists and doesn't get recreated with any extras.

Extras are needed to pass the URL from a notification to the
webview for navigation.