Changed notifications to account for missing tool
If a tool (such as sakai.announcement) is missing from the tools list
there was no default used as a fallback. This caused the notifications
list to wait for X notifcations while it only would receive X - 1
valid responses. The notification should fall back to the site if the
announcements tool is no longer available.

Also added the updated design colors for notifications.